Switch frigate to app-template
Stop using the frigate OG helm chart and use the standard app-template
one used in the rest of our apps.

For safety this will be rolled out manually, as it needs some work to
back up the frigate config and database and move them to their new
locations.

Both teslops and deltasite are moved to explicit PVCs for control rather
than the auto-provisioned PVs so the backup and move can be done in one
go